Pull results into own Results component
