+++ src/client/App.js
@@ -1,39 +1,18 @@
 import React, { Component } from 'react';
-import { connect } from 'react-redux';
 import Header from './Header';
 import Search from './search/Search';
-import Feed from './feed/Feed'
-import MostUsedWords from './results/MostUsedWords';
-import { groupedCountWords, mostFrequentWords, mostActiveUsers, exampleTweets } from './tweetAnalysis';
+import Results from './Results.js';
 
-const App = ({state}) => {
-  return (
-		<div>
-			<Header />
+const App = ({ feed }) => (
+  <div>
+    <Header />
 
-			<Search />
+    <Search />
 
-      <div className="ui grid">
-        <div className="four wide column">
-          <h3>Most frequent users</h3>
-        </div>
+    <Results />
 
-        <div className="eight wide column">
-          <Feed />
-        </div>
+    {/* <Footer /> *//* <Footer /> */}
+  </div>
+);
 
-        <div className="four wide column">
-          <MostUsedWords wordInfoList={groupedCountWords(mostFrequentWords(state.feed.map((post) => post.data)))} search=''/>
-        </div>
-      </div>
-
-			{/* <Footer /> */}
-		</div>
-	);
-};
-
-const mapStateToProps = (state) => {
-  return {state};
-};
-
-export default connect(mapStateToProps)(App);
+export default App;
